cloud_app/BuildingDamageProtection/requirements.txt
ibm_db==3.2.3
Flask==3.0.3
gevent==24.2.1
flask_restful==0.3.10
flask_httpauth==4.8.0
requests==2.31.0
paho-mqtt==1.6.1
pandas==2.2.0
pystache==0.6.5
matplotlib==3.8.2
gunicorn==21.2.0
